requirements.txt
pytz
Django==1.10
Pillow==3.3.0
requests==2.10.0
